requirements.txt
certifi==2021.5.30
dnspython==1.16.0
pymongo==3.12.0
selenium==3.141.0
urllib3==1.26.7
python-dotenv==0.19.1